Improve retry behaviour
Problem: When we tried to stream many operations, we found the
incremental backoff behaviour was not sufficient to alleviate rate-limit
problems.

Solution: Always wait *at least* as long as the rate-limit headers
specify, as well as increasing the number of retries and the rate at
which we increase the delay between requests.

Note: It turned out that a proxy we had set up between us and Contentful
was also enforcing its own very strict rate-limit. So the incremental
backoff is still important in the absence of rate-limit headers.