Don't crash when using --dry-run on an upload
Problem: Crashes with a JSON encoding error if attempting to --dry-run a
an operation that uploads a document (e.g. put-entry, post-entry).

Solution: Binary data needs decoded before putting it into a JSON
document. While it *should* be UTF-8, we don't enforce that, so decode
it using the "surrogateescape" scheme to be on the safe-side.