Be explicit about special environment behaviour
Problem: The previous approach to handling environment-aware resources,
like entries, content_types and assets, obfuscated where and when the
environment ID would be inserted into the request URL, hiding it in the
space_id field. This could confuse anyone trying to maintain the code.

Solution: Add "environment_path" as a special path fragment avaiable in
the template when the environment-aware flag is enabled on an endpoint.
This will be "/spaces/{space_id}" when no environment is specified, and
"/spaces/{space_id}/environments/{environment_id}" if an environment is
specified.